Allow published kernels to be patched AND configured
BUGZID: 10981
CQID:

With the previous commit for this bug/feature we were able to
add patches to a cloned kernel, but not have any related
configuration files processed. This is done by processing the
meta series that was created for the addon patches in the
configure phase. To not clobber the existing configuration,
I altered the behaviour when --continue is passed to kconf ..
to indicate that you are continuing and in progress kernel
configuration.

This commit also corrects the invalid starting point I created
when I imported a non-existent patch for feature/initramfs.

I also silenced the "already on branch" message in set_context,
since the multiple config now highlighted it on the screen.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>