merge broadcom_bcm91x80 BSP
CQID: WIND00128586

Enable the bcm91x80 be/le BSPs. Note that this configuration is
sharing common config items between the be and le board for the
first time in a WR Linux release. The be version of the board holds
the common patches and config items, the le version of the board
includes the common parts and adds little endian mods on top.

Higher layer boards should include the common feature as well to
get the patches and config. After that they should include the
kernel type and endianess they desire.

Signed-off-by: Weiwei Wang <Weiwei.wang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>