Kernel configuration changes for the emer_atca7101 BSP
Updated the existing mot_atca7101.cfg kernel config fragment
to work in 3.0. This involved minor changes as most of the options
that existed in 2.6.21 still exist in 2.6.27. A small change was
made to the edac.cfg as the main kernel config option was only a
choice between YES or NO not module. Renamed the BSP to emer_ as
well.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>