import fastboot changes from Ingo Molnar's linx-2.6-tip tree
CQID: 121345

This series of patches represents and effort to improve boot times
for x86 targets (and mainly x86 for these iterations). The series
is mainly the work of Arjan van de Ven <arjan@linux.intel.com> and
is part of the small kernel type.

Built and boot tested on several x86 targets during integration.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>