Remove errornous carriage return from IPMI serial protocol
    BUGZID:
    CQID:     WIND00135659

0013-ipmi-serial-terminal-mode-crlf-fix.patch erronously inserted
a carriage return in the middle of IPMI commands sent to the IPMC.
The fix this patch was attempting to accomplish (the insertion of
a CR at the end of IPMI commands) is no longer necessary because it
has already been applied to the upstream serial IPMI terminal mode
code.  As such, patch 0013, which was carried forward from wrlinux-2.0,
is no longer needed.