Import the 2.6.27 -rt kernel patch series
CQID: 121347

The -rt patch series is divided into two parts, the stock -rt
and the wrlinux "preempt_rt" kernel. The former applies directly
on top of a kernel.org kernel and gets you basic functionality.
preempt_rt is the applicationof that same rt series with the
rest of the Wind River Linux kernel functionality applied.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>