kernel config: migrate shared settings to shared config
We'd added a wrs_base config fragment with the intention
of putting the settings that were common across all kernel
types in that file.  This does that.

Now the settings that are shared across all kernel types
are clearly visible, and a sanity audit of others that can
be candidates for additional sharing is easier to do.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>