Add kernel-cache infrastructure for ftrace,early-ftrace and debugfs
CQID: WIND00122917

early
-----

early is a static tracer of early boottime functions.  It is used to profile
the early bootup of the kernel.

To configure, choose 'Early Tracing' in xconfig or manually set throudh the
config variable CONFIG_TRACE_EARLY.  Be aware that this is a static tracer
and will not function if CONFIG_DYNAMIC_FTRACE is set.

early produces its output when the system boots up.  Its output is accessed
through the normal ftrace method of reading the tracing/trace file in the
debugfs.  The tracer is started as early in the bootup as possible.  It
is started after kmem_cache_init in start_kernel.  This is an example of
output from x86.

          <idle>-0     [00]     0.001000: hpet_time_init <-start_kernel
	  <idle>-0     [00]     0.001000: hpet_enable <-hpet_time_init
	  <idle>-0     [00]     0.001000: setup_pit_timer <-hpet_time_init
	  <idle>-0     [00]     0.001000: time_init_hook <-hpet_time_init
	  <idle>-0     [00]     0.001368: pidmap_init <-start_kernel

Note that the timestamp does not start at 0.  There will be variation because
in this case, the tracing started after the system clock.  In some cases the
tracing will start before the system clock.  This is an example of the omap3430
start up.

          <idle>-0     [00]     0.000000: calibrate_delay <-start_kernel
	  <idle>-0     [00]     0.000000: pidmap_init <-start_kernel
	  <idle>-0     [00]     0.000000: anon_vma_init <-start_kernel
	  <idle>-0     [00]     0.000000: thread_info_cache_init <-start_kernel
- removed output -
          <...>-1     [00]     0.000000: omap_register_i2c_bus <-omap_i2c_init
          <...>-1     [00]     0.000000: omap_init_clocksource_32k <-__exception_text_end
          <...>-1     [00]  1433.023041: omap_init_devices <-__exception_text_end
          <...>-1     [00]  1433.024933: omap_init_dma <-__exception_text_end
Note that the timestamp for omap_init_devices looks valid of the call to
omap_init_clocksource_32k.

The functions listed are either in the kernel's init section, which is
unloaded when the kernel boots, or one of the many initcalls that are
responsible for initizing devices and kernel subsystems.  The tracing
continues until before the kernel hands off processing to the init process.

This is an example of the last part of a x86 trace log.
									                 <...>-1     [00]     3.858983: initrd_load <-prepare_namespace
         <...>-1     [00]     3.859173: rd_load_image <-initrd_load
         <...>-1     [00]     3.859456: mount_root <-prepare_namespace
         <...>-1     [00]     3.859458: nfs_root_data <-mount_root
         <...>-1     [00]     3.859463: root_nfs_parse <-nfs_root_data
         <...>-1     [00]     3.859465: root_nfs_parse <-nfs_root_data
         <...>-1     [00]     3.931151: do_mount_root <-mount_root