rt: remove 26rt8 content before replacement.
CQID: 121347

The RT patchset is going to use the filenames as per the series
file used "upstream" in order to give a consistent basis for
maintenance and future bugfixes.

Since there is a refresh in between, the git rename logic doesn't
do anything useful for us here either, so move these out of the way
first.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>