preempt_rt: update kernel type based on rt feature
CQID: 121347

With the rt feature updated, the preempt_rt kernel type needs
its scc file and patch transforms updated to match.  The main
difference is the handling of the <trace/sched.h> file that
was originally in the ftrace-upstream.patch of RT but really
originally parented from the LTT feature.  Since the ktype here
is a blend of RT and LTT, we have to drop the file addition from
RT, let LTT add it, and then patch in the RT specific tracepoint
addition.  All the rest was basic context related stuff.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>