preempt_irq: remove old patch files.
We are moving towards having the patches as named per what is in the
preempt_rt series files, vs. the old files which had the patch names
as per what "git format-patch" created.  Since there is a refresh
involved, the rename detection doesn't do much for us, so just remove
the old ones.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>