preemptible irq: bring in the basline patches for 2.6.27
CQID: 121346

This brings in the patches for preemptible IRQ, the subset
of the RT patches.  To be consistent with what we do with RT,
we'll have the baseline patches in "irq" and then the context
mods and so forth to make it work on a WR standard kernel in
"preempt_irq".

Note that in 2.0, all the patches were in "preempt_irq"

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>