preempt_irq: Create the wrapper feature around the mainline
The feature "irq" contains the patches as per what would apply
to a default 2.6.27 kernel.  This wrapper contains the context
mods that we need to deploy in order to get these same patches
to apply to the WR standard kernel.  The wrapper feature is
called "preempt_irq" in order to be consistent with 2.0 naming.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>