update to 2.6.27.5 and fix context diffs
CQID: 143344

The 2.6.27.5 stable update was released. This updates the
base of the tree to that version. The .5 update had acpi
updates, which required both the common_pc and -rt patches
to be modified slightly. -rt is just a context update and
common_pc gets to drop 3 patches since they made the
-stable update.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>