fix CFS group fair scheduling in 2.6.27
CQID: 144162

The linux kernel mailing list saw several reports of
group fair scheduling not working perfectly for groups
that contained a single task. The following to commits
are taking from the kernel.org git repos to address
the problem.

Build and boot tested on x86 and powerpc architectures.

The following upstream commits map to these patches:

 38736f47 -> 0001-sched-fix-__load_balance_iterator-for-cfq-with-on.patch
 caea8a03 -> 0002-sched-fix-list-traversal-to-use-_rcu-variant.patch

Details about the changes are in the headers of the patches
themselves.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>