skip cop2 register save to fix kernel thread stack corruption
CQID: WIND00144232

Code in "resume" can corrupt the stack of kernel thread. It assumes
there is a frame of "pt_regs" at the bottom of the kernel stack. But
if it's in a kernel thread, there is no such a frame at the bottom
of the stack and the instructions that update stored cp0_status
corrupt the stack.

Then following instructions in "resume" corrupt the stack of kernel
thread:

=======================================================
       /* Disable COP2 in the stored process state */
       li      t1, ST0_CU2
       xor     t0, t1
       LONG_S  t0, ST_OFF(t2)
=======================================================

The fix is to judge whether it's in kernel thread by value of
"addr_limit" in "struct thread_info". If value of "addr_limit"
is equal to "KERNEL_DS", it's in a kernel thread and we don't
save cop2 registers.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>