create FUSE feature kernel config options
CQID: 144999

The design of 3.0 states:
that whenever possible kernel options should be
in the kernel-cache (so we can find them all), but
can be enabled via a template.

To do this for FUSE, we move the feature template
from wrll-wrlinux -> wrll-linux-2.6.27, create a
linux subdir and a fuse.scc description. We also
create a features/fuse.scc in the kernel-cache that
simply turn FUSE on as a module.

The wrll-linux-2.6.27 feature template description
simply includes the kernel-cache feature to get
the associated kernel option

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>