netns: Coexist with the sysfs limitations v2
CQID: 146428

To make testing of the network namespace simpler allow
the network namespace code and the sysfs code to be
compiled and run at the same time.  To do this only
virtual devices are allowed in the additional network
namespaces and those virtual devices are not placed
in the kobject tree.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>