Add interface to disable NMI interrupt from user-space at runtime.
CQID: 131898

The disabling of NMI is done by reverting the line on which NMIs
come in on the local APIC to a regular interrupt line, and then
masking it. Since this is not normal operating behaviour, errors
must be masked to prevent them from being recorded by the kernel.

The functionality is provided for both 32-bit and 64-bit kernels.

The interface is provided via the /proc/sys/kernel/nmi_enabled
entry. Writing a 0 to it disables NMI interrupts, writing a 1
re-enables them.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>