EDAC: fix edac device rmmod deadlock
CQID: WIND00143185

The work of an edac device has to be completed before the whole edac_dev
structure being removed. flush_cpu_workqueue() is used to wait for all
works in edac_poller's workqueue to be processed, by inserting a wq_barrier
into the tail of the workqueue, then sleeping on the completion of this
wq_barrier, expecting edac_poller to run wq_barrier.work to wake it up.

Works of all current edac devices share the same callback function of
edac_device_workq_function(), which would grab the mutex of
device_ctls_mutex before it checks the device. This is exactly where edac_poller and rmmod
would have a great chance to deadlock each other - on one hand rmmod would
sleep on the completion of a wq_barrier, holding device_ctls_mutex; on the
other hand edac_poller would be blocked on the same mutex when it's running
any one of works of existing edac evices(Note, this edac_dev.work is likely
to be totally irrelevant to the one that is being removed right now)and never
would have a chance to run the work of above wq_barrier to wake rmmod up.

That's why edac_device_workq_teardown() should be moved outside of the critical
region of device_ctls_mutex. Moreover, an edac_dev.work should bail out
immediately if it is being removed.

Signed-off-by: Harry Ciao<qingtao.cao@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>