ppc64: re-enable some more patches from the original RT queue
CQID: 121346

Two patches in the preempt_rt series are responsible for changing
the scheduling points called from entry_64 on ppc64; but due to
the poor choice of names given by the original author, they were
originally overlooked when selecting patches relating directly to
preemptible IRQ support.  There is also a third, that impacts
xmon; while not as critical, the fix is valid in preempt_irq
context and not just preempt_rt context.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>