ipi call: use raw lock.
    CQID: 121347

preempt-realtime-x86_64.patch used to bump the call_lock up to a
raw lock, however 3b16cf8 removed the lock;  implicit in the change,
is that call_function_lock (kernel/smp.c) is now in charge of keeping
things sane.  For sanity to remain, we also need to bump up this
one to a raw lock.  (see also b7d7a2404).

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>