preempt-realtime-net.patch: clear STATE_SCHED before qdisc_run()
CQID: 121346

If you look at the applied patch against a 2.6.26 kernel, you
see it takes the lock and does the qdisc_run() -- both in the
original code path, and in the added codepath for SOFTIRQS.

However, in moving to 2.6.27, the memory barrier and the
clear_bit() were added in the default code path from upstream
changes.  Of course patch didn't know to update the new (but
basically cloned) RT path to match.  This fixes that oversight.

Since this is for softirq, it is required here in the preempt_irq
feature (or, irq baseline feature to be more correct).

Signed-off-by: Liming Wang <liming.wang@windriver.com>
Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>