fsl_mpc8360e add USB host controller support
CQID: WIND00146436

The following series of 7 patches are intended to add USB Host
Controller support.

First 4 patches are backported from the community and the last 3
are copied from the linux-usb mail list.

They still are not applied to the tree but but already passed several
spins in the maillist.

You can find these patches here:
http://git.kernel.org/?p=linux/kernel/git/paulus/powerpc.git;a=commit;h=64b60e096fa391c56f93e6216115e6757bf86b7e
http://patchwork.ozlabs.org/patch/11982/
http://patchwork.ozlabs.org/patch/11984/
http://patchwork.ozlabs.org/patch/11985/
http://article.gmane.org/gmane.linux.kernel/736679
http://article.gmane.org/gmane.linux.usb.general/10289
http://article.gmane.org/gmane.linux.usb.general/10290

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>