CF fsl_mpc5121e_ads (USB OTG PM bug fix)
CQID: WIND00128494

In wrlinux-2.0 (linux-2.6.21.7), before system enters into sleep
or deep sleep the kernel has the capability to let these devices
which are not already in suspend status enter into suspend status,
however, in wrlinux-3.0. kernel PM subsystem has lost this capability,
thus cause some drivers' suspend function to be reentrant, in this case
that is to say, the kernel whole PM mechanism is not consistent with
USB OTG's, so needs to synchronize them.

The patch modifies the USB PM functions which are dedicated to mpc5121e
ADS board to address the issue as above stated, however, because USB on
mpc5121e ADS boards uses big endian registers and big endian data description
which are rather different from USB on other powerpc boards, so they can't be
shared.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>