EDAC: release vMC sensor LUN number
CQID: WIND00149273

EDAC module should release the LUN number it used during removal,
this way it could have a chance to pick up the same number if it's
re-installed again. The benefit of doing so is that if there are more
than one EDAC modules and one of them gets removed and then
re-installed,
it's expected that they could still get different LUN numbers.

Signed-off-by: Harry Ciao <qingtao.cao@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>