kernel/signal.o failed on sun_niagara2_sun4v-ecgl_glibc-cgl
CQID: WIND00150317

When we merged nested-alt-sigstack.patch, we forgot that
sparc is building the -cgl kernels. As a result, signal.c
is failing to compile, since ECGL is enabling the feature
but asm/signal.h doesn't define the limit for sparc.

We could have forced the feature off for sparc, but that's
not the best thing to do. The define is simple, so we add
it for sparc to get the kernel building.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>