rt12: incorporate the new patches added for v2.6.26.8-rt12
CQID: WIND00150547

This brings in the added patches that 1st appeared in rt12 (released
during the holidays).  Note that the choice of patch name
"rtmutex-tglx-is-a-moron.patch" is in upstream, and has nothing to
do with me.

We were fortunate that someone had an SLA that meant tglx is updating
the 2.6.24 RT kernel, since that meant the patches were carried over
onto 2.6.26, and then we could make the final jump to our 2.6.27.

Given that proper handling of rtmutex is core to having preempt_rt
behave itself, and that tglx is probably one of the few people to
really understand things in this area, we would be silly to not
integrate these bugfixes from him.

None of the other already existing patches in rt11 were changed or
deleted in the move to rt12.  Integrating the new rt12 patches onto
our 2.6.27 kernel was straightforward, so there should be no merge
induced errors either.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>