Cavium-aware gdb fails to connect to cav_ebt5800 target
CQID: WIND00150587

[
The hooks to allow the Cavium-aware gdb debugger to connect to the
target were missing. This primarily involved registration of an
interrupt handler to pulse the MCD0 line on all cores to place them in
the debug state when a character is received on the debug serial
port.

Conditional code to add this hook and the mechanism for enabling the
debugger at boot time through the use of a boot command line parameter
were added.

Inclusion of the hook is selected when the kernel is configured by
selecting the CONFIG_CAVIUM_GDB option and ensuring that
CONFIG_CAVIUM_WATCHDOG is NOT selected. It is no longer necessary to
turn off CONFIG_KGDB or to reduce the number of serial ports supported
in the kernel from 2 to 1 when configuring with
CONFIG_CAVIUM_GDB. However, when the kernel is booted with the
'cavium_gdb' parameter, use of the second serial port (/dev/ttyS1) for
normal communication purposes is disabled. The selection of serial
port (/dev/ttyS1) is currently fixed.

Furthermore, if both the 'cavium_gdb' and 'debug' parameters are
specified on the boot command line, the boot will be halted in the
early stages to wait for the debugger to be connected before
continuing the boot process. This allows debugging of the boot
process, similar to the use of the 'kgdbwait' commandline option to
KGDBoC.
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>