edac: Add missing error detections for AMD8131 chipset
CQID: WIND00146330
[
As what we have agreed, once we use edac_pci structures to present the 4
PCI-X controllers on 6101, their sysfs director
ies would be name as "pciX", where "X" is integer of 0~3, determined by
each of their edac_pci structure's internal inde
x. So, "ctl_name" string is freed and we could put more descriptive
information into it as what it is supposed to be - t
he name of the specific controller.

From below logs we could tell that "ctl_name" is used to format debug
information or error messages, having nothing to d
o with EDAC sysfs directory name.
]

folded into: 0029-EDAC-add-new-vMC-sensors-for-AMD8131.patch
and added: 0029-EDAC-add-new-vMC-sensors-for-AMD8131.patch

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>