preempt_irq: remove more patches that aren't strictly needed
CQID: WIND00150699

These are patches that are not necessary for the implementation
of preempt_irq.

Testing was initiated on a 2.6.26.8 kernel, by starting with the full RT
series of ~450 patches (but with RT off), and then adding several
patches that were required to make preempt_IRQ stable there.  Then that
patch set was boiled down to the remaining ~60 in some 22 steps, in
order to be able to track functionality along the way so that
regressions would be immediately noticed.  Testing was on SMP ppc32
(sbc8641d), SMP x86 (intel mpcbl0001), SMP x86-64 (wallowa) and SMP
ppc64 (pasemi).

Full details of the 22 steps and testing can be found here:

http://yow-lpdgit.ottawa.wrs.com/gitweb/gitweb.cgi?p=local/pgortmak/IRQ-26/linux-2.6;a=summary

http://yow-lpdgit.ottawa.wrs.com/gitweb/gitweb.cgi?p=local/pgortmak/IRQ-26/patches;a=summary

Once the stable and minimal working set was determined, this was used to
further refine the existing 2.6.27 based work.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>