preempt_irq: refresh of the context mods required for CGL.
The "irq" patch series is that which applies to the default
kernel.org 2.6.27 tree -- the preempt_irq feature is a wrapper
around that with context modes to allow it to work with the
WR CGL kernel. This is a refresh of those context mods to
go with the updates to the irq patchset.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>