net traffic control: coalesce options in one place
CQID: 152169

If we are going to have a feature that dictates our support
on net traffic control, then all the related option settings for
that feature should come from that feature and not from the
parent kernel types that include it.

Currently it only had the following options:

 CONFIG_NET_CLS=y
 CONFIG_NET_SCH_HTB=y
 CONFIG_NET_CLS_CGROUP=y
 CONFIG_NET_CLS_BASIC=y

and the rest (some two dozen or so) were repeated in the various
kernel types.  This removes them from the kernel types and puts
them all in one spot, here in the feature.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>