net traffic control: move options from base config to feature
CQID: 152169

Similar to what was done for the config options set in the kernel
types in the prev commit, this moves the feature related options
out from the base config fragment and into the domain of the feature
where they belong.

Even if the feature ended up having no patches (in some future release),
the idea of having the related config options all isolated in one frag
would make it easier to exclude them all in one go.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>