enumerate LTTng system call descriptions on MIPS, ARM, PPC
CQID: WIND00150552

LTTng for ARM/MIPS/PPC is missing HAVE_LTT_DUMP_TABLES. Without this
feature, traces with syscalls only show the syscall number, the name of the
syscall is not reported. With this patch applied, the syscall ID and the name of
the syscall handler is reported.

But for ppc64, the pair of syscalls, the normal 64bit and 32bit-compatible,
share the same syscall id, and lttv cannot distinguish them. So the logs
will only show the latter one registered, in this patch, it's the
32bit-compatible syscall name.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>