Add a Kconfig for oom-ignore-child-vmsize
CQID: WIND00151289

[
The old patch removed the lines completely.

This patch wraps them in #ifndef CONFIG_OOM_IGNORE_CHILD Kconfig.
This Kconfig is off by default which keeps the code intact.

According to the ECGL Feature migration document:
"Do not account the vmsize of a child when calculating the parent
process' badness value: Merge with CGL, make it a configurable kernel option
is disabled by default."
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>