Fix page faults, incorrect syscall traces in oprofile
CQID: WIND00151335

[

The patch replaces the system call frame detection logic to correct the
discovery of system call frames when backtracing in the kernel.  The
original algorithm assumed that if the sp was not a valid kernel stack pointer
and the instruction pointer was within bounds of certain system call
handlers, then a valid system call was found.  This did not take account for other
entry methods into the entry_xx.S common syscall code.

....

]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>