preempt_irq: absorb the relevant changes from rt13
CQID: 150699

The upstream preempt_rt series for .26 kernel was updated from 12 to
13 and contains one patch removal (not used for .27 anyway) and two
new additions.  One of the new additions is relevant to preempt_irq;
it impacts drivers that use IRQF_RANDOM, and possible warnings that
they may emit.  This brings in the patch and resyncs the series file
(in this case the scc file) to be in line with the rt series file
for ease of tracking changes.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>