Make detection for faulty 8250 Serial UARTs optional
CQID: WIND00132192

[
This patch adds a new Kconfig option which may be used to disable the
UART_BUG_TXEN auto-detection feature.

This feature from mainline is intended to detect UARTs that don't
properly raise an interrupt when the transmitter buffer is empty.
It can occassionally give false positives which interferes with
IPMI messages causing a race condition between serial8250_interrupt
and serial8250_start_tx.

This new option defaults to on to match the mainline kernel.  This
patch turns it off for the boards with 8250-compatible UARTs.

The targets were booted with the option disabled and IPMI messages
were successfully exchanged with the IPMC through the ipmitool.
The serial console output also functioned properly without incident.

Back history:  This was a fix for a critical problem observed in a customer's
lab.  The UART was incorrect being detected as faulty once in a hundred
or so reboots.  This wrecked havok on the IPMI serial driver that was using
the serial 8250.  Both interrupt-based events and polling were being
performed simulataneously.

]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>