cgroups: readjust configs for small kernels.
CQID: 152169

As a followup to commit 1b8b617, I had the action to use the
automated build output to audit the resulting configs.  For
small kernels, this change will put several config settings
back to "off" that were changed to "=m" as a result of 1b8b617.
Now the small kernels ahve the same settings as they did
prior to 1b8b617.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>