kconfig: silence silly warnings from new HOSTCC on kernel build
CQID: 152169

While testing the cgroup config changes, I noticed that a modern
HOSTCC (like that of Ubuntu 8.10) generates lots of silly warnings
about unused return values from stdio functions, and about missing
initializer values from the output of gperf. (When I say "lots",
it can be in excess of twenty...)

Since these show up right at the beginning of a kernel config or
at the beginning of a kernel compile, they have rather high
visibility, and significantly diminish the confidence in the end
product, even though these are nothing to do with what we've done.
This is a purely cosmetic patchset to mask the cheezy warnings, it
makes no effort to do any clever checks on the otherwise ignored
return values.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>