Add performance counter support to the NEC EMMA 3P BSP
CQID: WIND00152838

[
There was no driver for the CPU variant found on the EMMA so following
the example of the op_model_mipsxx.c file I created one for the VR5500 CPU
variant. This is not a pretty driver and can use some cleanup but it
seems to work and I had Dave Lerner check it out as well to compare with what
he sees on other targets and all is good. checkpatch.pl will give an
error due to the macro in vr5500_perfcount_handler() but my quick attempts to
fix this didn't get me anywhere.

So basically this is known not to be perfect but does seem to work well.

Notice the EMMA report the cpu as R9721 but from all I can find this is
a VR5500.
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>