ecgl: fix Kconfig options
CQID: WIND00151111

This is combination of Kconfig options that need to be set for CGL/ECGL.

The CONFIG_PMEM and CONFIG_PMEMFS options should be set in feature/pmem,
however the individual BSP PMEM options should be in the bsp
configuration files.  I've placed each in <bsp>-cgl.cfg since the PMEM feature is
included in CGL by default.  Each piece of hardware is auto-detected on
each board anyway so if the hardware is not present it falls back to a
simulated PMEM mode.

I created an ECGL config file for the emer_atca6101, 7101 and 717.
They all contain CONFIG_SERIAL_8250_BUG_TXEN=n which disables auto-detection
of faulty 8250 UARTS.  I also enabled CONFIG_CPC925_ECC_SCRUB for the
6101.

"#CONFIG_QUOTA is not set" was removed from the ECGL config file at
Joe's request.

CONFIG_QDISC_STATS was also added to CGL to match the qdisc.patch file
moving to CGL.

CONFIG_OOM_IGNORE_CHILD was enabled for ECGL.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>