CGL kernel does not boot with >= 64GB and NUMA
CQID: WIND00152005

[
The Netra-T5440 kernel boot death was traced to the
CONFIG_PAX_MEMORY_SANITIZE feature.  If there's 64GB or more memory
installed, the kernel gets stuck in __free_pages_bootmem() when trying
to free memory on the second NUMA node.  Disabling this option allows the
kernel to boot, and all 64GB of memory is recognized.

It is not clear if CONFIG_PAX_MEMORY_SANITIZE has problems with NUMA
*and* SPARC64, or if it has problems with NUMA and large memory in
general.

There are no other targets available with NUMA (namely dual Opteron
x86-64) and at least 64GB of memory installed to test this out.

To err on the side of caution, I opted to make
CONFIG_PAX_MEMORY_SANITIZE incompatible with CONFIG_NUMA (so this
fix applies to all architectures).
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>