fix bsdjail
CQID: WIND00153521

[
The core of preventing jailbreak in the bsdjail module was commented
out to make it compile.  I uncommented it back in, and dealt with the
compilation issue, namely adding a call to lookup the dentry from an
inode.  I also added some extra debug statements which can be enabled
through the regular debug mechanism of the module.

Further testing showed that the proc filesystem was not accessible
from within the jail, and caused a soft lockup of the processor.  The
problem turned out to be a missing advancement of a variable in the
path of the lsm callout in the base of the proc filesystem.  Other
similar paths nearby in the code had that variable advancement.
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>