Cleanup config options that were causing warnings
CQID:WIND00154108

While working on another issue related to these BSPs serveral
warnings were being issues indicated that the following config
options are no longer recognized. As such they are being removed.

In the case of CONFIG_USB_EHCI_BIG_ENDIAN_MMIO is still valid but not
for the BSP that was attempting to use it based on the required
dependencies.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>