Add code to handle powerpc serial interrupt flood
CQID:WIND00154108

[
In 2.0 Paul added code to get around an issue with newer FSL powerpc
SoC based targets getting an interrupt flood when a break was issued
on the serial console (for example when issuing a magic Sysrq). This
is a software fix for what is most likely a hardware issue. These changes
are almost the same as what was using in 2.0 however the enabling of the
fix has been moved to use a single kernel config value rather then
swithing on or off based on the board.
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>