preempt_rt: tlb cleanup, more dangerous than you could imagine.
CQID: 150713

Upstream cleanup 42b777281 (remove double indirection **tlb)
exposes a subtle issue in this patch; that it depended semi
implictly on it being a double in this line:

	*tlb = tlb_gather_mmu(vma->vm_mm, fullmm);

Once tlb becomes of local scope only, (from 42b777) then you
get things unravelling in nasty ways in exit_mm() and the like.

I'd tested by creating a local:

	struct mmu_gather *mytlb = *tlb;

and I'd used that everywhere, except for the interchange of these two
lines, indicated as + in this pseudo patch:

-               *tlb = tlb_gather_mmu(vma->vm_mm, fullmm);
+               mytlb = tlb_gather_mmu(vma->vm_mm, fullmm);
+//             mytlb = *tlb = tlb_gather_mmu(vma->vm_mm, fullmm);

of which I'd only use one or the other, as a completely isolated change.

The currently commented out line (equiv. to orig code) works; the
use of the local that sheilds **tlb from a change causes the boot
to fail miserably.

Since this patch was originally targeting a latency observed in 2.6.15,
it is not clear that it is even still of value ; the newly born git
tree of RT patches does not have it anymore.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>