cgl: increase log buf size for dmesg.
CGL boards tend to have a lot of PCI bridges and bootup messages,
which will wrap the dmesg ring buffer when left to its default
size.  Move the option setting it to 14 from base to std, and
add an option for cgl to up it to 15.

Note that the option has moved from "Kernel hacking" into the
"General Setup" category since the 2.6.21 kernel.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>