cramfs: Fix non-xip cramfs flash boot failure
CQID: 152424

[
	From Tiejun Chen:

This issue is already fixed by me when developing BSP fsl_mpc8349eimtx.
As you know this BSP will be not release for 3.0 so this issue still
exit on 3.0 currently. But looks this problem should one issue based on
the cramfs fro all BSPs so I recommend Liang submit quickly the patch
after he validate that on his BSP. Otherwise the cramfs is not supported
correctly for 3.0.

The issue results from there is no i_fop function used to get correct
inode information for cramfs. And that error is generated from one
feature patch: xip/cramfs-linear-xip.patch

]

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>